requirements.txt
torch
torchvision
albumentations
scikit-learn
matplotlib
seaborn
Pillow
tqdm
streamlit
onnxruntime
torchxrayvision
scikit-image
